names fixed, constant sound switched to one that doesn't cause popping (#76)
Changed move to active for naming. The audio popping appeared to be caused by the enemy constant sound being played along with others, replacing it with different sounds seems to fix the issue so it's just a dog growl for now until we start really deciding on sounds. Also converted everything to .wav format since that's best for a desktop game, and removed all the unnecessary unused sound files we were accumulating.