fix enemy light detection bug and move shared function to AIController (#84)
Fixes #82. Issue was AIController was checking player coordinates which after PR #79, are no longer descriptive of player light center.

Also refactored helped function checking if enemy is within player light radius to AIController bc it is shared across mult AI types and to facilitate this I moved the playermodel handling to AIController.

This (having light centered on player) does raise a gameplay issue where if you are walking down, because the player is so close to the bottom of the light radius, you can die basically instantly.